+++ src-tauri/src/settings.rs
@@ -53,6 +53,10 @@ fn get_default_settings() -> Value {
         },
         "tradingCards": {
             "priceAdjustment": 0.0,
+            "sellLimit": {
+                "min": 0.01,
+                "max": 1.10
+            }
         },
     })
 }